+++ hutool-cache/src/main/java/com/xiaoleilu/hutool/cache/impl/FIFOCache.java
@@ -35,6 +35,10 @@ public class FIFOCache<K, V> extends AbstractCache<K, V> {
 	 * @param timeout 过期时长
 	 */
 	public FIFOCache(int capacity, long timeout) {
+		if(Integer.MAX_VALUE == capacity) {
+			capacity -= 1;
+		}
+		
 		this.capacity = capacity;
 		this.timeout = timeout;
 		cacheMap = new LinkedHashMap<K, CacheObj<K, V>>(capacity + 1, 1.0f, false);